ai_character_chat_ver1.1/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pillow==10.3.0
python-multipart==0.0.9
google-generativeai==0.7.0
